Refactor JsonSchema module into focused converter modules
Major refactoring to improve maintainability and organization:

**Module Split:**
- Extract FromSkema converter (Skema → JSON Schema) - 268 lines
- Extract ToSkema converter (JSON Schema → Skema) - 194 lines
- Streamline main JsonSchema module to API only - 143 lines (76% reduction)

**Test Split:**
- Split converter-specific tests into focused test files
- Clean up main test file to integration tests only - 164 lines (84% reduction)
- Maintain complete test coverage with 173 tests

**New Feature:**
- Add :pattern as alias for :format in Skema schemas
- Support both regex and string patterns
- :format takes precedence when both present

**Benefits:**
- Improved separation of concerns
- Better maintainability and readability
- Focused test organization
- Parallel development capability
